brackup-restore's verbose flag is more verbose now, showing files as they're restored.
brackup-restore can restore from an encrypted *.brackup file now,
firing up gpg for user to decrypt to a tempfile



git-svn-id: http://brackup.googlecode.com/svn/trunk@153 c3152786-d43a-0410-8499-09eb43983aed

+++ lib/Brackup/Util.pm
@@ -5,16 +5,18 @@ require Exporter;
 
 use vars qw(@ISA @EXPORT_OK);
 @ISA = ('Exporter');
-@EXPORT_OK = ('tempfile', 'tempdir');
+@EXPORT_OK = qw(tempfile tempdir slurp valid_params);
 
 my $mainpid = $$;
 my @TEMP_FILES = ();  # ([filename, caller], ...)
 
 END {
     # will happen after File::Temp's cleanup
-    foreach my $rec (@TEMP_FILES) {
-        next unless -e $rec->[0];
-        unlink($rec->[0]);
+    if ($$ == $mainpid) {
+        foreach my $rec (@TEMP_FILES) {
+            next unless -e $rec->[0];
+            unlink($rec->[0]);
+        }
     }
 }
 use File::Temp ();
@@ -22,12 +24,26 @@ use File::Temp ();
 sub tempfile {
     my (@ret) = File::Temp::tempfile();
     my $from = join(" ", (caller())[0..2]);
-
+    push @TEMP_FILES, [$ret[1], $from];
     return wantarray ? @ret : $ret[0];
 }
 
 sub tempdir {
     return File::Temp::tempdir(@_);
+}
+
+sub slurp {
+    my $file = shift;
+    open(my $fh, $file) or die "Failed to open $file: $!\n";
+    return do { local $/; <$fh>; }
+}
+
+sub valid_params {
+    my ($vlist, %uarg) = @_;
+    my %ret;
+    $ret{$_} = delete $uarg{$_} foreach @$vlist;
+    croak("Bogus options: " . join(', ', sort keys %uarg)) if %uarg;
+    return %ret;
 }
 
 1;
